+++ server/core/src/projects/index.ts
@@ -1,3 +1,4 @@
 export { addProject } from './addProject'
 export { getProjects } from './getProjects'
 export { deleteProject } from './deleteProject'
+export { updateProject } from './updateProject'