+++ web/src/routes/+layout.svelte
@@ -3,7 +3,7 @@
   import Header from '$lib/components/Header.svelte'
 </script>
 
-<div class="min-h-screen bg-black flex flex-col">
+<div class="min-h-screen primary-bg flex flex-col">
   <Header />
   <main>
     <slot />